Create plottingmean.m
This function can be used to plot the average of a plotted figure without the original x and y data. Original intent of the code was to plot the average line of several lines of unequal length. The function takes a figure as input and produces the average plot, average x and y values as output.